+++ NEWS.md
@@ -1,11 +1,18 @@
-# Changes in version 1.0.2 (release date:??-??-2024)
+# Changes in version 1.0.2 (release date:26-10-2024)
 
 - Added a `NEWS.md` file to track changes to the package.
 - Stops interactive calling of `chooseCRANmirror` on `.onAttach` if interactive and CRAN mirror not set GGIR #1141.
+- GGIRread version look-up in .onattach() no longer crashes when computer is offline.
+- Migrate read function for ActiGraph (csv) and Actiwatch (csv/awd) count data to GGIRread #68.
+- Add function for reading Actical (csv) count data #68.
+- Add functions for reading and merging Philips Health Band file pairs (xlsx) #68.
+- Add functions for reading and merging Fitbit (json) files with sleep, steps, and/or calories #68.
+
 
 # Changes in version 1.0.1 (release date:03-06-2024)
 
 - Progress bar fixed, issue #63 (credits: John Muschelli)
+
 
 # Changes in version 1.0.0 (release date:27-03-2024)